ProductoPolicies

+++ app/Http/Controllers/ProductoController.php
@@ -26,6 +26,7 @@ class ProductoController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny',Producto::class);
         $productos = Producto::with("user")->simplePaginate(10);
         return view("productos.index",compact("productos"));
     }
@@ -37,6 +38,7 @@ class ProductoController extends Controller
      */
     public function create()
     {
+        $this->authorize('create',Producto::class);
         $producto=new Producto;
         $title=__("Crear producto");
         $textButton=__("Crear");
@@ -54,6 +56,8 @@ class ProductoController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create',Producto::class);
+
         $this->validate($request, [
             "nombre" => "required|max:50|unique:productos",
             "precio"=>"required",
@@ -74,6 +78,8 @@ class ProductoController extends Controller
      */
     public function show(Producto $producto)
     {
+        $this->authorize('viewAny',Producto::class);
+
         $title=__("Mostrar producto");
         $textButton=__("Regresar");
         $route=route("productos.index");
@@ -88,6 +94,7 @@ class ProductoController extends Controller
      */
     public function edit(Producto $producto)
     {
+        $this->authorize('update',Producto::class);
         $update= true;
         $title=__("Editar producto");
         $textButton=__("Actualizar");
@@ -106,6 +113,7 @@ class ProductoController extends Controller
      */
     public function update(Request $request, Producto $producto)
     {
+        $this->authorize('update',Producto::class);
         $this->validate($request, [
             "nombre" => "required|max:50|unique:productos,nombre," . $producto->id,
             "precio"=>"required",
@@ -126,6 +134,7 @@ class ProductoController extends Controller
      */
     public function destroy(Producto $producto)
     {
+        $this->authorize('delete',Producto::class);
         if(Gate::allows('admin')){
             $producto->delete();
             return back()->with("success", __("¡Producto eliminado!"));
